requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via azure-functions-durable
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via aiohttp
azure-appconfiguration==1.7.2
    # via
    #   -r requirements.in
    #   azure-appconfiguration-provider
azure-appconfiguration-provider==2.1.0
    # via -r requirements.in
azure-core==1.38.1
    # via
    #   azure-appconfiguration
    #   azure-appconfiguration-provider
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
azure-functions==1.21.3
    # via
    #   -r requirements.in
    #   azure-functions-durable
azure-functions-durable==1.2.10
    # via -r requirements.in
azure-identity==1.25.2
    # via -r requirements.in
azure-keyvault-secrets==4.10.0
    # via azure-appconfiguration-provider
azure-storage-blob==12.25.1
    # via -r requirements.in
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==46.0.5
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via azure-appconfiguration-provider
et-xmlfile==2.0.0
    # via openpyxl
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.4
    # via azure-functions-durable
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.7.2
    # via
    #   azure-appconfiguration
    #   azure-keyvault-secrets
    #   azure-storage-blob
jiter==0.13.0
    # via openai
msal==1.34.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
numpy==2.4.2
    # via -r requirements.in
openai==1.69.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
orderedmultidict==1.0.2
    # via furl
polars==1.38.1
    # via -r requirements.in
polars-runtime-32==1.38.1
    # via polars
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via msal
python-dateutil==2.9.0.post0
    # via azure-functions-durable
regex==2026.1.15
    # via tiktoken
requests==2.32.5
    # via
    #   azure-core
    #   azure-functions-durable
    #   msal
    #   tiktoken
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sniffio==1.3.1
    # via openai
tenacity==9.1.4
    # via -r requirements.in
tiktoken==0.12.0
    # via -r requirements.in
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   azure-appconfiguration
    #   azure-core
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
yarl==1.22.0
    # via aiohttp
